CI: Make Divingbell Ubuntu Job non-voting
The Divingbell Ubuntu job is currently a voting job, even though it does
not run as a gate job on merged changes. Since the OpenDev migration,
the job has failed consistently. New issues have surfaced while
attempting to correct the issue [0].

This change moves the job to non-voting in order to allow the merge of
critical work while the cause of the failures is investigated.

[0] https://review.opendev.org/663392

Change-Id: Ie44bb7f0160acc362af212028cc553ec99090acd
Signed-off-by: Drew Walters <andrew.walters@att.com>