perm: Various fixes (values hash, revert)
The hash used by divingbell-perms to decide whether or not to rerun the
permissions script was being generated incorrectly, using a fixed value
instead of actually looking at the values passed to the chart.

This change updates the hash to reflect conf.divingbell.perms, and will
rerun the script if the hash changes.

Also fixes the logic to revert permissions.

Change-Id: I74f056f69a1b7f0eb9223915b1671e1e18091483